requirements.txt
requests
beautifulsoup4
anthropic
openai
playwright
schedule
feedparser
python-dateutil
pyyaml
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
